// apps/web/app/api/generated/schemas/getProjectsId200.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Catalyst API
 * Type-safe REST API for the Catalyst project management system
 * OpenAPI spec version: 1.0.0
 */
import type { GetProjectsId200Description } from './getProjectsId200Description';
import type { GetProjectsId200Status } from './getProjectsId200Status';

export type GetProjectsId200 = {
  id: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  description?: GetProjectsId200Description;
  status: GetProjectsId200Status;
  createdAt: string;
  updatedAt: string;
};